# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Yuni-Q",
    "description": "Make Node.js",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "mighty-gorge-25136.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/posts": {
      "get": {
        "summary": "게시판 전체 호출",
        "tags": [
          "posts"
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      },
      "post": {
        "summary": "게시글 작성",
        "tags": [
          "posts"
        ],
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": {
            "type": "object",
            "properties": {
              "todo": {
                "type": "string",
                "example": "aa"
              },
              "date": {
                "type": "date",
                "example": "2018-10-30"
              }
            }
          }
        }],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      }
    },
...
    "/posts/{_id}/edit": {
      "get": {
        "summary": "게시판 수정 페이지 호출",
        "tags": [
          "posts"
        ],
        "parameters": [{
          "name": "_id",
          "in": "path",
          "type": "string",
          "default": "aaaaa"
        }],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      }
    },
    "/posts/{_id}": {
      "get": {
        "summary": "게시글 수정",
        "tags": [
          "posts"
        ],
        "parameters": [{
          "name": "_id",
          "in": "path",
          "type": "string",
          "default": "aaaaa"
        }],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      },
      "put": {
        "summary": "게시글 수정",
        "tags": [
          "posts"
        ],
        "parameters": [{
            "name": "_id",
            "in": "path",
            "type": "string",
            "default": "aaaaa"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "todo": {
                  "type": "string",
                  "example": "aa"
                },
                "date": {
                  "type": "date",
                  "example": "2018-10-30"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/Post"
            }
          }
        }
      },
      "delete": {
        "summary": "게시글 삭제",
        "tags": [
          "posts"
        ],
        "parameters": [{
          "name": "_id",
          "in": "path",
          "type": "string",
          "default": "aaaaa"
        }],
        "responses": {
          "200": {}
        }
      }
    },
...
  },
  "definitions": {
...
    "Post": {
      "required": [
        "todo",
        "date"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "todo": {
          "type": "string",
          "uniqueItems": true
        },
        "date": {
          "type": "date",
          "uniqueItems": true
        },
        "userId": {
          "type": "string"
        }
      }
    }
  }
}